Move registry code into repo.
The experimental BOSH registry code from
github.com/frodenas/bosh-registry is moved into the registry package in
this repository. Subsequent commits will refactor it into a GCE-specefic
implementation that reads registry settings from the GCE metadata
service as opposed to the BOSH registry.

[#116463251](https://www.pivotaltracker.com/story/show/116463251)
