+++ src/bosh-google-cpi/action/attach_disk.go
@@ -7,7 +7,7 @@ import (
 	"bosh-google-cpi/google/disk_service"
 	"bosh-google-cpi/google/instance_service"
 
-	"github.com/frodenas/bosh-registry/client"
+	"bosh-google-cpi/registry"
 )
 
 type AttachDisk struct {
