+++ src/bosh-google-cpi/action/concrete_factory.go
@@ -19,7 +19,7 @@ import (
 	"bosh-google-cpi/google/subnetwork_service"
 	"bosh-google-cpi/google/target_pool_service"
 
-	"github.com/frodenas/bosh-registry/client"
+	"bosh-google-cpi/registry"
 )
 
 type ConcreteFactory struct {
