+++ src/bosh-google-cpi/action/concrete_factory_options.go
@@ -3,7 +3,7 @@ package action
 import (
 	bosherr "github.com/cloudfoundry/bosh-utils/errors"
 
-	"github.com/frodenas/bosh-registry/client"
+	"bosh-google-cpi/registry"
 )
 
 type ConcreteFactoryOptions struct {
