+++ src/bosh-google-cpi/action/detach_disk_test.go
@@ -9,9 +9,9 @@ import (
 	. "bosh-google-cpi/action"
 
 	instancefakes "bosh-google-cpi/google/instance_service/fakes"
-	registryfakes "github.com/frodenas/bosh-registry/client/fakes"
+	registryfakes "bosh-google-cpi/registry/fakes"
 
-	"github.com/frodenas/bosh-registry/client"
+	"bosh-google-cpi/registry"
 )
 
 var _ = Describe("DetachDisk", func() {
