+++ src/bosh-google-cpi/registry/agent_options.go
