+++ src/bosh-google-cpi/registry/agent_settings.go
