+++ src/bosh-google-cpi/registry/client.go
