+++ src/bosh-google-cpi/registry/client_options.go
