+++ src/bosh-google-cpi/registry/doc.go
