+++ src/bosh-google-cpi/registry/http_client.go
